Secondary check in PHP when deleting booking to make sure cutoff hasn't passed
Fixes #15

+++ actions/booking_delete.php
@@ -3,6 +3,12 @@ include_once("../inc/autoload.php");
 
 $bookingObject = new booking($_POST['bookingUID']);
 
-$bookingObject->delete();
-
+$mealObject = new meal($bookingObject->meal_uid);
+if (date('Y-m-d H:i:s') >= date('Y-m-d H:i:s', strtotime($mealObject->date_cutoff))) {
+  $bookingObject->delete();
+} else {
+  if ($_SESSION['admin'] == true) {
+    $bookingObject->delete();
+  }
+}
 ?>
